Move automaton intersection to automata.py
